@remotion/studio: Revert canvas outline migration
