+++ packages/canvas/src/canvas-controller.ts
@@ -3,10 +3,6 @@ import type {TSequence} from 'remotion';
 import {calculateTimeline} from './calculate-timeline';
 import type {TimelineTrackData} from './get-timeline-sequence-sort-key';
 import {
-	createCanvasOutlinesController,
-	type CanvasOutlinesController,
-} from './outlines';
-import {
 	createCanvasSelectionController,
 	type CanvasSelectionController,
 } from './selection';
@@ -16,7 +12,6 @@ export type CanvasController = {
 		readonly getSnapshot: () => readonly TimelineTrackData[];
 		readonly subscribe: (listener: () => void) => () => void;
 	};
-	readonly outlines: CanvasOutlinesController;
 	readonly selection: CanvasSelectionController;
 };
 
@@ -51,7 +46,6 @@ export const createCanvasController = (): CanvasController => {
 				return () => timelineListeners.delete(listener);
 			},
 		},
-		outlines: createCanvasOutlinesController(),
 		selection: createCanvasSelectionController(),
 	};
 
@@ -64,10 +58,7 @@ export const createCanvasController = (): CanvasController => {
 				}),
 			);
 		},
-		clear: () => {
-			controller.outlines.disconnect();
-			updateTimelineSnapshot([]);
-		},
+		clear: () => updateTimelineSnapshot([]),
 	});
 
 	return controller;
