+++ packages/canvas/tsconfig.json
@@ -7,7 +7,6 @@
 		"target": "ES2022",
 		"module": "es2020",
 		"moduleResolution": "bundler",
-		"allowJs": true,
 		"declaration": true,
 		"emitDeclarationOnly": true
 	},
