+++ packages/studio/src/components/SelectedOutlineKeyboardControls.tsx
@@ -1,4 +1,3 @@
-import type {CanvasSelectableOutline} from '@remotion/canvas';
 import {PlayerInternals} from '@remotion/player';
 import type React from 'react';
 import {useCallback, useContext, useEffect, useRef} from 'react';
@@ -16,6 +15,7 @@ import {
 	type SelectedOutlineKeyframedDragChange,
 	type SelectedOutlineStaticDragChange,
 } from './selected-outline-drag';
+import type {getSequencesWithSelectableOutlines} from './selected-outline-measurement';
 import {
 	getSelectedSequenceKeys,
 	getSequenceKeysContainingSelection,
@@ -35,7 +35,9 @@ export const SelectedOutlineKeyboardControls: React.FC<{
 	readonly getLatestOutlineTargetByKey: (
 		key: string,
 	) => SelectedOutlineTarget | undefined;
-	readonly getSelectableOutlines: () => readonly CanvasSelectableOutline[];
+	readonly getSelectableOutlines: () => ReturnType<
+		typeof getSequencesWithSelectableOutlines
+	>;
 }> = ({getLatestOutlineTargetByKey, getSelectableOutlines}) => {
 	const currentSelection = useCurrentTimelineSelectionStateAsRef();
 	const {getDragOverrides} = useContext(
