+++ packages/studio/src/components/selected-outline-measurement.ts
@@ -1,6 +1,15 @@
+import type {_InternalTypes, OverrideIdToNodePaths, TSequence} from 'remotion';
+import {calculateTimeline} from '../helpers/calculate-timeline';
 import {BLACK, WHITE} from '../helpers/colors';
+import {useCache as resetBoxQuadsCache} from '../helpers/get-box-quads-polyfill-internals.js';
+import {getBoxQuadsPonyfill} from '../helpers/get-box-quads-ponyfill';
+import {timelineSequenceNodePathToKey} from '../helpers/timeline-node-path-key';
 import type {OutlinePoint, SelectedOutline} from './selected-outline-geometry';
 import {mixPoint} from './selected-outline-geometry';
+import type {
+	SelectedOutlineLayoutTarget,
+	SequenceWithSelectedOutline,
+} from './selected-outline-types';
 import {
 	cropFieldKeys,
 	rotateFieldKey,
@@ -163,6 +172,184 @@ export const getSelectedOutlineRotationPivot = ({
 	}
 
 	return getUvHandlePosition(points, uv);
+};
+
+const rectToPoints = (
+	elementRect: DOMRect,
+	containerRect: DOMRect,
+): SelectedOutline['points'] => {
+	const left = elementRect.left - containerRect.left;
+	const top = elementRect.top - containerRect.top;
+	const right = elementRect.right - containerRect.left;
+	const bottom = elementRect.bottom - containerRect.top;
+
+	return [
+		{x: left, y: top},
+		{x: right, y: top},
+		{x: right, y: bottom},
+		{x: left, y: bottom},
+	];
+};
+
+type SvgViewport = {
+	readonly x: number;
+	readonly y: number;
+	readonly width: number;
+	readonly height: number;
+};
+
+type SvgScreenCtm = Pick<DOMMatrixReadOnly, 'a' | 'b' | 'c' | 'd' | 'e' | 'f'>;
+
+export const getTransformedSvgViewportPoints = ({
+	viewport,
+	ctm,
+	containerRect,
+}: {
+	readonly viewport: SvgViewport;
+	readonly ctm: SvgScreenCtm;
+	readonly containerRect: Pick<DOMRect, 'left' | 'top'>;
+}): SelectedOutline['points'] => {
+	const transformPoint = (x: number, y: number): OutlinePoint => ({
+		x: ctm.a * x + ctm.c * y + ctm.e - containerRect.left,
+		y: ctm.b * x + ctm.d * y + ctm.f - containerRect.top,
+	});
+
+	const left = viewport.x;
+	const top = viewport.y;
+	const right = viewport.x + viewport.width;
+	const bottom = viewport.y + viewport.height;
+
+	return [
+		transformPoint(left, top),
+		transformPoint(right, top),
+		transformPoint(right, bottom),
+		transformPoint(left, bottom),
+	];
+};
+
+const quadToPoints = (
+	quad: DOMQuad,
+	containerRect: DOMRect,
+): SelectedOutline['points'] => {
+	// `getBoxQuads`/the ponyfill returns the quad in viewport coordinates.
+	// The overlay <svg> is unscaled (the canvas `scale()`/pan live on a sibling
+	// container, not the svg), so 1 user unit == 1 px and we only need to move
+	// the quad into the svg's local space by subtracting its viewport origin.
+	// We deliberately do not pass `relativeTo` to the ponyfill: when the target
+	// is not an ancestor of the element, the polyfill cannot resolve the
+	// coordinate space and leaves the quad in viewport coordinates.
+	return [
+		{x: quad.p1.x - containerRect.left, y: quad.p1.y - containerRect.top},
+		{x: quad.p2.x - containerRect.left, y: quad.p2.y - containerRect.top},
+		{x: quad.p3.x - containerRect.left, y: quad.p3.y - containerRect.top},
+		{x: quad.p4.x - containerRect.left, y: quad.p4.y - containerRect.top},
+	];
+};
+
+const isSvgSvgElement = (element: Element): element is SVGSVGElement => {
+	const ownerSvgSvgElement = element.ownerDocument.defaultView?.SVGSVGElement;
+	return (
+		(typeof SVGSVGElement !== 'undefined' &&
+			element instanceof SVGSVGElement) ||
+		(ownerSvgSvgElement !== undefined && element instanceof ownerSvgSvgElement)
+	);
+};
+
+const getSvgSvgElementViewport = (element: SVGSVGElement): SvgViewport => {
+	const viewBox = element.viewBox.baseVal;
+	if (viewBox.width > 0 && viewBox.height > 0) {
+		return {
+			x: viewBox.x,
+			y: viewBox.y,
+			width: viewBox.width,
+			height: viewBox.height,
+		};
+	}
+
+	return {
+		x: 0,
+		y: 0,
+		width: element.width.baseVal.value,
+		height: element.height.baseVal.value,
+	};
+};
+
+const getSvgSvgElementOutlinePoints = (
+	element: SVGSVGElement,
+	containerRect: DOMRect,
+): SelectedOutline['points'] | null => {
+	const ctm = element.getScreenCTM();
+	const viewport = getSvgSvgElementViewport(element);
+	if (ctm === null || (viewport.width === 0 && viewport.height === 0)) {
+		return null;
+	}
+
+	return getTransformedSvgViewportPoints({
+		viewport,
+		ctm,
+		containerRect,
+	});
+};
+
+const getElementOutlinePoints = (
+	element: Element,
+	containerRect: DOMRect,
+	includeOutsideContainer: boolean,
+): SelectedOutline['points'] | null => {
+	const elementRect = element.getBoundingClientRect();
+
+	if (elementRect.width === 0 && elementRect.height === 0) {
+		return null;
+	}
+
+	if (
+		!includeOutsideContainer &&
+		(elementRect.right <= containerRect.left ||
+			elementRect.left >= containerRect.right ||
+			elementRect.bottom <= containerRect.top ||
+			elementRect.top >= containerRect.bottom)
+	) {
+		return null;
+	}
+
+	if (isSvgSvgElement(element)) {
+		return getSvgSvgElementOutlinePoints(element, containerRect);
+	}
+
+	const quads = getBoxQuadsPonyfill(element, {
+		box: 'border',
+	});
+	const quad = quads?.[0];
+	if (!quad) {
+		return rectToPoints(elementRect, containerRect);
+	}
+
+	return quadToPoints(quad, containerRect);
+};
+
+export const cropOutlinePoints = (
+	points: SelectedOutline['points'],
+	crop: SelectedOutlineLayoutTarget['crop'],
+): SelectedOutline['points'] => {
+	if (
+		crop.left === 0 &&
+		crop.right === 0 &&
+		crop.top === 0 &&
+		crop.bottom === 0
+	) {
+		return points;
+	}
+
+	const {left, top} = crop;
+	const right = 1 - crop.right;
+	const bottom = 1 - crop.bottom;
+
+	return [
+		getUvHandlePosition(points, [left, top]),
+		getUvHandlePosition(points, [right, top]),
+		getUvHandlePosition(points, [right, bottom]),
+		getUvHandlePosition(points, [left, bottom]),
+	];
 };
 
 export const getSelectedSequenceKeys = (
@@ -408,4 +595,172 @@ export const getSelectedRotationInfo = (
 				? selectedItem.frame
 				: selectedItem.fromFrame,
 	};
+};
+
+export const getSequencesWithSelectableOutlines = ({
+	sequences,
+	overrideIdsToNodePaths,
+	compositions = [],
+	timelinePosition,
+}: {
+	readonly sequences: readonly TSequence[];
+	readonly overrideIdsToNodePaths: OverrideIdToNodePaths;
+	readonly compositions?: readonly _InternalTypes['AnyComposition'][];
+	readonly timelinePosition: number;
+}): SequenceWithSelectedOutline[] => {
+	return calculateTimeline({
+		sequences: [...sequences],
+		overrideIdsToNodePaths,
+		compositions,
+	})
+		.filter((track) => {
+			if (track.nodePathInfo === null) {
+				return false;
+			}
+
+			return (
+				track.sequence.showInTimeline &&
+				timelinePosition >= track.sequence.from &&
+				timelinePosition < track.sequence.from + track.sequence.duration &&
+				track.nodePathInfo.auxiliaryKeys.length === 0
+			);
+		})
+		.filter((track) => track.sequence.refForOutline !== null)
+		.sort((a, b) => a.depth - b.depth)
+		.map((track) => {
+			if (track.nodePathInfo === null) {
+				throw new Error('Expected selected outline to have a node path');
+			}
+
+			return {
+				depth: track.depth,
+				keyframeDisplayOffset: track.keyframeDisplayOffset,
+				key: getTimelineSequenceSelectionKey(track.nodePathInfo),
+				nodePathInfo: track.nodePathInfo,
+				sequence: track.sequence,
+			};
+		});
+};
+
+export const measureOutlines = (
+	container: SVGSVGElement,
+	targets: readonly SelectedOutlineLayoutTarget[],
+	hoveredTimelineNodePathKey: string | null,
+): SelectedOutline[] => {
+	return measureOutlineTargets(
+		container,
+		targets.map((target) => ({
+			crop: target.crop,
+			includeOutsideContainer:
+				target.showSelectedOutline ||
+				timelineSequenceNodePathToKey(
+					target.nodePathInfo.sequenceSubscriptionKey,
+				) === hoveredTimelineNodePathKey,
+			key: target.key,
+			ref: target.ref,
+		})),
+	);
+};
+
+export const measureOutlineTargets = (
+	container: Element,
+	targets: readonly {
+		readonly crop: SelectedOutlineLayoutTarget['crop'];
+		readonly includeOutsideContainer: boolean;
+		readonly key: string;
+		readonly ref: React.RefObject<Element | null>;
+	}[],
+): SelectedOutline[] => {
+	// Reuse shared ancestor geometry within this synchronous batch only.
+	resetBoxQuadsCache();
+	const containerRect = container.getBoundingClientRect();
+	const outlines: SelectedOutline[] = [];
+
+	for (const target of targets) {
+		const element = target.ref.current;
+		if (element === null) {
+			continue;
+		}
+
+		const uncroppedPoints = getElementOutlinePoints(
+			element,
+			containerRect,
+			target.includeOutsideContainer,
+		);
+		if (uncroppedPoints === null) {
+			continue;
+		}
+
+		const points = cropOutlinePoints(uncroppedPoints, target.crop);
+
+		outlines.push({
+			key: target.key,
+			dimensions:
+				element instanceof HTMLElement
+					? {
+							width: element.offsetWidth,
+							height: element.offsetHeight,
+						}
+					: element instanceof SVGSVGElement
+						? {
+								width: element.width.baseVal.value,
+								height: element.height.baseVal.value,
+							}
+						: null,
+			uncroppedPoints,
+			points,
+		});
+	}
+
+	return outlines;
+};
+
+export const outlinesAreEqual = (
+	a: readonly SelectedOutline[],
+	b: readonly SelectedOutline[],
+): boolean => {
+	if (a.length !== b.length) {
+		return false;
+	}
+
+	for (let i = 0; i < a.length; i++) {
+		if (a[i].key !== b[i].key) {
+			return false;
+		}
+
+		if (
+			a[i].dimensions?.width !== b[i].dimensions?.width ||
+			a[i].dimensions?.height !== b[i].dimensions?.height
+		) {
+			return false;
+		}
+
+		const aUncropped = a[i].uncroppedPoints;
+		const bUncropped = b[i].uncroppedPoints;
+		if ((aUncropped === null) !== (bUncropped === null)) {
+			return false;
+		}
+
+		if (aUncropped !== null && bUncropped !== null) {
+			for (let j = 0; j < aUncropped.length; j++) {
+				if (
+					Math.abs(aUncropped[j].x - bUncropped[j].x) > 0.01 ||
+					Math.abs(aUncropped[j].y - bUncropped[j].y) > 0.01
+				) {
+					return false;
+				}
+			}
+		}
+
+		for (let j = 0; j < a[i].points.length; j++) {
+			if (
+				Math.abs(a[i].points[j].x - b[i].points[j].x) > 0.01 ||
+				Math.abs(a[i].points[j].y - b[i].points[j].y) > 0.01
+			) {
+				return false;
+			}
+		}
+	}
+
+	return true;
 };
