+++ packages/studio/src/components/selected-outline-order.ts
@@ -1,17 +1,5 @@
-import type {CanvasOutline} from './outlines';
-
-export type CanvasOutlineOrderTarget = {
-	readonly key: string;
-	readonly containsSelection: boolean;
-	readonly selected: boolean;
-	readonly sequence: {
-		readonly id: string;
-		readonly parent: string | null;
-	};
-};
-
-type SelectedOutline = CanvasOutline;
-type SelectedOutlineLayoutTarget = CanvasOutlineOrderTarget;
+import type {SelectedOutline} from './selected-outline-geometry';
+import type {SelectedOutlineLayoutTarget} from './selected-outline-types';
 
 const outlinePointEqualityTolerance = 0.5;
 const outlineAreaEqualityTolerance = 0.5;
@@ -365,7 +353,7 @@ const orderOutlineGroup = ({
 	return ordered;
 };
 
-export const orderCanvasOutlinesForRendering = ({
+export const orderOutlinesForRendering = ({
 	outlines,
 	sequences,
 	targetsByKey,
