+++ packages/studio/src/components/selected-outline-types.ts
@@ -1,4 +1,3 @@
-import type {CanvasOutlineTarget} from '@remotion/canvas';
 import type {
 	CanUpdateSequencePropStatus,
 	CanUpdateSequencePropStatusKeyframed,
@@ -20,11 +19,12 @@ export type SelectedOutlineContextMenuOpenHandler = () =>
 	| SelectedOutlineContextMenuOpenResult
 	| Promise<SelectedOutlineContextMenuOpenResult>;
 
-export type SelectedOutlineLayoutTarget = CanvasOutlineTarget & {
+export type SelectedOutlineLayoutTarget = {
 	readonly key: string;
 	readonly containsSelection: boolean;
 	readonly keyframeDisplayOffset: number;
 	readonly nodePathInfo: SequenceNodePathInfo;
+	readonly ref: React.RefObject<Element | null>;
 	readonly selected: boolean;
 	readonly selectedForCrop: boolean;
 	readonly selectedForRotation: boolean;
@@ -34,6 +34,12 @@ export type SelectedOutlineLayoutTarget = CanvasOutlineTarget & {
 	readonly transformOriginValue: string;
 	readonly selection: TimelineSelection;
 	readonly sequence: TSequence;
+	readonly crop: {
+		readonly left: number;
+		readonly right: number;
+		readonly top: number;
+		readonly bottom: number;
+	};
 };
 
 export type SelectedOutlineTarget = SelectedOutlineLayoutTarget & {
@@ -216,6 +222,14 @@ export type SelectedOutlineRotationDragState = {
 	readonly startRotation: readonly [number, number, number];
 	readonly startValue: string;
 	readonly target: SelectedOutlineRotationDragTarget;
+};
+
+export type SequenceWithSelectedOutline = {
+	readonly depth: number;
+	readonly keyframeDisplayOffset: number;
+	readonly key: string;
+	readonly nodePathInfo: SequenceNodePathInfo;
+	readonly sequence: TSequence;
 };
 
 export const translateFieldKey = 'style.translate';
