+++ packages/studio/src/components/selected-outline-uv.ts
@@ -1,7 +1,3 @@
-import {
-	getCanvasOutlinePointAtUv,
-	getCanvasOutlineUvForPoint,
-} from '@remotion/canvas';
 import type {
 	CanUpdateSequencePropStatusKeyframed,
 	CanUpdateSequencePropStatusStatic,
@@ -15,6 +11,8 @@ import type {
 import {Internals} from 'remotion';
 import {
 	clamp,
+	mix,
+	mixPoint,
 	type OutlinePoint,
 	type SelectedOutline,
 } from './selected-outline-geometry';
@@ -272,10 +270,84 @@ export const tuplesEqual = (left: unknown, right: UvCoordinate): boolean => {
 	return left[0] === right[0] && left[1] === right[1];
 };
 
+const getBilinearUvHandlePosition = (
+	points: SelectedOutline['points'],
+	uv: UvCoordinate,
+): OutlinePoint => {
+	const [tl, tr, br, bl] = points;
+	const top = mixPoint(tl, tr, uv[0]);
+	const bottom = mixPoint(bl, br, uv[0]);
+	return mixPoint(top, bottom, uv[1]);
+};
+
+type ProjectiveTransform = {
+	readonly a: number;
+	readonly b: number;
+	readonly c: number;
+	readonly d: number;
+	readonly e: number;
+	readonly f: number;
+	readonly g: number;
+	readonly h: number;
+};
+
+const projectiveEpsilon = 0.000001;
+
+const getProjectiveTransform = (
+	points: SelectedOutline['points'],
+): ProjectiveTransform | null => {
+	const [tl, tr, br, bl] = points;
+	const dx1 = tr.x - br.x;
+	const dx2 = bl.x - br.x;
+	const dx3 = tl.x - tr.x + br.x - bl.x;
+	const dy1 = tr.y - br.y;
+	const dy2 = bl.y - br.y;
+	const dy3 = tl.y - tr.y + br.y - bl.y;
+
+	let g = 0;
+	let h = 0;
+	if (Math.abs(dx3) > projectiveEpsilon || Math.abs(dy3) > projectiveEpsilon) {
+		const determinant = dx1 * dy2 - dx2 * dy1;
+		if (Math.abs(determinant) < projectiveEpsilon) {
+			return null;
+		}
+
+		g = (dx3 * dy2 - dx2 * dy3) / determinant;
+		h = (dx1 * dy3 - dx3 * dy1) / determinant;
+	}
+
+	return {
+		a: tr.x - tl.x + g * tr.x,
+		b: bl.x - tl.x + h * bl.x,
+		c: tl.x,
+		d: tr.y - tl.y + g * tr.y,
+		e: bl.y - tl.y + h * bl.y,
+		f: tl.y,
+		g,
+		h,
+	};
+};
+
+const applyProjectiveTransform = (
+	transform: ProjectiveTransform,
+	uv: UvCoordinate,
+): OutlinePoint => {
+	const denominator = transform.g * uv[0] + transform.h * uv[1] + 1;
+	return {
+		x: (transform.a * uv[0] + transform.b * uv[1] + transform.c) / denominator,
+		y: (transform.d * uv[0] + transform.e * uv[1] + transform.f) / denominator,
+	};
+};
+
 export const getUvHandlePosition = (
 	points: SelectedOutline['points'],
 	uv: UvCoordinate,
-): OutlinePoint => getCanvasOutlinePointAtUv(points, uv);
+): OutlinePoint => {
+	const transform = getProjectiveTransform(points);
+	return transform === null
+		? getBilinearUvHandlePosition(points, uv)
+		: applyProjectiveTransform(transform, uv);
+};
 
 export const getUvHandleConnectionLines = ({
 	handles,
@@ -534,10 +606,80 @@ export const getUvEllipseInteractiveControls = ({
 	return controls;
 };
 
+const vectorBetween = (from: OutlinePoint, to: OutlinePoint): OutlinePoint => {
+	return {x: to.x - from.x, y: to.y - from.y};
+};
+
+const getBilinearUvCoordinateForPoint = (
+	points: SelectedOutline['points'],
+	point: OutlinePoint,
+): UvCoordinate => {
+	const [tl, tr, br, bl] = points;
+	let u = 0.5;
+	let v = 0.5;
+
+	for (let i = 0; i < 8; i++) {
+		const current = getBilinearUvHandlePosition(points, [u, v]);
+		const errorX = current.x - point.x;
+		const errorY = current.y - point.y;
+		if (Math.abs(errorX) + Math.abs(errorY) < 0.001) {
+			break;
+		}
+
+		const du = {
+			x: mix(tr.x - tl.x, br.x - bl.x, v),
+			y: mix(tr.y - tl.y, br.y - bl.y, v),
+		};
+		const dv = vectorBetween(mixPoint(tl, tr, u), mixPoint(bl, br, u));
+		const determinant = du.x * dv.y - du.y * dv.x;
+		if (Math.abs(determinant) < 0.000001) {
+			break;
+		}
+
+		u -= (errorX * dv.y - errorY * dv.x) / determinant;
+		v -= (du.x * errorY - du.y * errorX) / determinant;
+	}
+
+	return [u, v];
+};
+
 export const getUvCoordinateForPoint = (
 	points: SelectedOutline['points'],
 	point: OutlinePoint,
-): UvCoordinate => getCanvasOutlineUvForPoint(points, point);
+): UvCoordinate => {
+	const transform = getProjectiveTransform(points);
+	if (transform === null) {
+		return getBilinearUvCoordinateForPoint(points, point);
+	}
+
+	const determinant =
+		transform.a * (transform.e - transform.f * transform.h) -
+		transform.b * (transform.d - transform.f * transform.g) +
+		transform.c * (transform.d * transform.h - transform.e * transform.g);
+	if (Math.abs(determinant) < projectiveEpsilon) {
+		return getBilinearUvCoordinateForPoint(points, point);
+	}
+
+	const inverseA = transform.e - transform.f * transform.h;
+	const inverseB = transform.c * transform.h - transform.b;
+	const inverseC = transform.b * transform.f - transform.c * transform.e;
+	const inverseD = transform.f * transform.g - transform.d;
+	const inverseE = transform.a - transform.c * transform.g;
+	const inverseF = transform.c * transform.d - transform.a * transform.f;
+	const inverseG = transform.d * transform.h - transform.e * transform.g;
+	const inverseH = transform.b * transform.g - transform.a * transform.h;
+	const inverseI = transform.a * transform.e - transform.b * transform.d;
+
+	const denominator = inverseG * point.x + inverseH * point.y + inverseI;
+	if (Math.abs(denominator) < projectiveEpsilon) {
+		return getBilinearUvCoordinateForPoint(points, point);
+	}
+
+	return [
+		(inverseA * point.x + inverseB * point.y + inverseC) / denominator,
+		(inverseD * point.x + inverseE * point.y + inverseF) / denominator,
+	];
+};
 
 export function constrainUv(
 	value: UvCoordinate,
