+++ packages/studio/src/helpers/get-box-quads-polyfill-internals.js
